moved everything that is not pure OGM outside
