+++ docs/type.index.md
@@ -16,11 +16,7 @@ Indexes are entities which contain the following attributes:
 ## Examples
 
 ```php
-use Syndesi\CypherDataStructures\Type\IndexName;
-use Syndesi\CypherDataStructures\Type\Index;
-use Syndesi\CypherDataStructures\Type\IndexType;
-use Syndesi\CypherDataStructures\Type\OGM\NodeLabel;
-use Syndesi\CypherDataStructures\Type\PropertyName;
+use Syndesi\CypherDataStructures\Type\Index;use Syndesi\CypherDataStructures\Type\IndexName;use Syndesi\CypherDataStructures\Type\IndexType;use Syndesi\CypherDataStructures\Type\NodeLabel;use Syndesi\CypherDataStructures\Type\PropertyName;
 
 $index = new Index();
 $index
