+++ tests/Type/NodeLabelTest.php
@@ -6,7 +6,7 @@ namespace Syndesi\CypherDataStructures\Tests\Type;
 
 use PHPUnit\Framework\TestCase;
 use Syndesi\CypherDataStructures\Exception\InvalidArgumentException;
-use Syndesi\CypherDataStructures\Type\OGM\NodeLabel;
+use Syndesi\CypherDataStructures\Type\NodeLabel;
 
 class NodeLabelTest extends TestCase
 {
