+++ tests/Type/WeakRelationStorageTest.php
@@ -9,10 +9,10 @@ use SplObjectStorage;
 use stdClass;
 use Syndesi\CypherDataStructures\Exception\InvalidArgumentException;
 use Syndesi\CypherDataStructures\Exception\LogicException;
-use Syndesi\CypherDataStructures\Type\OGM\Relation;
-use Syndesi\CypherDataStructures\Type\OGM\RelationType;
-use Syndesi\CypherDataStructures\Type\OGM\WeakRelation;
-use Syndesi\CypherDataStructures\Type\OGM\WeakRelationStorage;
+use Syndesi\CypherDataStructures\Type\Relation;
+use Syndesi\CypherDataStructures\Type\RelationType;
+use Syndesi\CypherDataStructures\Type\WeakRelation;
+use Syndesi\CypherDataStructures\Type\WeakRelationStorage;
 
 class WeakRelationStorageTest extends TestCase
 {
